feat: complete Modules task
